Move metadata value types into a metadata package
Regenerates phonemetadata.proto into a new metadata package (mirroring
upstream's Phonemetadata) -- the only acyclic home for these types once the
loader depends on them. The four value types (PhoneMetadata, NumberFormat,
PhoneNumberDesc, PhoneMetadataCollection) are re-exported from the root package
as type aliases, so the public API is unchanged; their hand-written helpers
(Merge/HasPossibleLength/HasPossibleLengthLocalOnly) move with them and are now
exported. The metadata loader/source code stays in root for now (a later step).
Unblocks moving MatchNationalNumber into internal/ (#5b).
