+++ phonemetadata.proto
@@ -24,6 +24,7 @@ syntax = "proto2";
 option optimize_for = LITE_RUNTIME;
 
 option java_package = "com.google.i18n.phonenumbers";
+option go_package = "github.com/nyaruka/phonenumbers/v2/metadata;metadata";
 package phonenumbers;
 
 message NumberFormat {
