+++ shortnumberinfo.go
@@ -110,7 +110,7 @@ func IsPossibleShortNumber(number *PhoneNumber) bool {
 		if phoneMetadata == nil {
 			continue
 		}
-		if phoneMetadata.GeneralDesc.hasPossibleLength(int32(shortNumberLength)) {
+		if phoneMetadata.GeneralDesc.HasPossibleLength(int32(shortNumberLength)) {
 			return true
 		}
 	}
@@ -128,7 +128,7 @@ func IsPossibleShortNumberForRegion(number *PhoneNumber, regionDialingFrom strin
 		return false
 	}
 	numberLength := len(GetNationalSignificantNumber(number))
-	return phoneMetadata.GeneralDesc.hasPossibleLength(int32(numberLength))
+	return phoneMetadata.GeneralDesc.HasPossibleLength(int32(numberLength))
 }
 
 // Tests whether a short number matches a valid pattern. If a country calling code is shared by
@@ -185,7 +185,7 @@ func GetExpectedCostForRegion(number *PhoneNumber, regionDialingFrom string) Sho
 	// The possible lengths are not present for a particular sub-type if they match the general
 	// description; for this reason, we check the possible lengths against the general description
 	// first to allow an early exit if possible.
-	if !phoneMetadata.GetGeneralDesc().hasPossibleLength(int32(len(shortNumber))) {
+	if !phoneMetadata.GetGeneralDesc().HasPossibleLength(int32(len(shortNumber))) {
 		return UNKNOWN_COST
 	}
 
@@ -320,7 +320,7 @@ func matchesPossibleNumberAndNationalNumber(number string, numberDesc *PhoneNumb
 	if numberDesc == nil {
 		return false
 	}
-	if len(numberDesc.PossibleLength) > 0 && !numberDesc.hasPossibleLength(int32(len(number))) {
+	if len(numberDesc.PossibleLength) > 0 && !numberDesc.HasPossibleLength(int32(len(number))) {
 		return false
 	}
 	return MatchNationalNumber(number, numberDesc, false)
